+++ src/dice/mod.rs
@@ -7,6 +7,7 @@ mod roll;
 mod roll_result;
 
 use crate::redis::*;
+use serde::Serialize;
 use std::{fmt, str::FromStr};
 
 pub use die::*;