+++ src/interface/dice.rs
@@ -80,6 +80,19 @@ pub unsafe extern "C" fn roll_execute(ptr: *const Roll) -> *mut RollResult {
     Box::into_raw(Box::new(result))
 }
 
+/// Get the string representation of a roll
+/// # Safety
+/// Panics via assert! on a null ptr
+#[no_mangle]
+pub unsafe extern "C" fn roll_to_string(ptr: *const Roll) -> *mut c_char {
+    let roll = {
+        assert!(!ptr.is_null());
+        &*ptr // unsafe
+    };
+    let c_str = CString::new(roll.to_string()).unwrap();
+    c_str.into_raw()
+}
+
 //
 // RollResult
 //
@@ -106,6 +119,20 @@ pub unsafe extern "C" fn roll_result_free(ptr: *mut RollResult) {
         return;
     }
     Box::from_raw(ptr);
+}
+
+/// Get the JSON string from a RollResult
+/// # Safety
+/// Panics via assert! on a null ptr
+#[no_mangle]
+pub unsafe extern "C" fn roll_result_to_json(ptr: *const RollResult) -> *mut c_char {
+    let roll_result = {
+        assert!(!ptr.is_null());
+        &*ptr // unsafe
+    };
+    let json = serde_json::to_string(&roll_result).unwrap();
+    let c_str = CString::new(json).unwrap();
+    c_str.into_raw()
 }
 
 /// Get the string representation of a roll result