+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 I was premature in jumping to 0.1.  **This is the only working version of 0.1**.  Semantic version, take two!
 
 * Actually implement everything needed for the 0.1 demo.
+* Add serde, RollResult, Roll JSON interface
 
 # 0.1.1